+++ src/main/java/cn/omisheep/authz/core/util/JSONDecryptUtils.java
@@ -1,6 +1,6 @@
 package cn.omisheep.authz.core.util;
 
-import cn.omisheep.commons.util.RsaHelper;
+import cn.omisheep.commons.util.RSAHelper;
 import cn.omisheep.commons.util.StringUtils;
 import com.alibaba.fastjson.JSONObject;
 
@@ -12,20 +12,21 @@ import java.util.Arrays;
  */
 public class JSONDecryptUtils {
 
-    public static JSONObject decrypt(String name, JSONObject obj, String key) {
-        if (!StringUtils.hasText(name)) return null;
+    public static void decrypt(String name, JSONObject obj, String key) {
+        if (!StringUtils.hasText(name)) return;
         String[] trace = Arrays.stream(name.split("\\.")).distinct().toArray(String[]::new);
-        return decrypt(trace, obj, key);
+        decrypt(trace, obj, key);
     }
 
-    public static JSONObject decrypt(String[] trace, JSONObject obj, String key) {
-        if (obj == null) return null;
+    public static void decrypt(String[] trace, JSONObject obj, String key) {
+        if (obj == null) return;
 
         if (trace.length == 1) {
-            obj.put(trace[0], RsaHelper.decrypt(obj.get(trace[0]).toString(), key));
-            return obj;
+            if (obj.get(trace[0]) instanceof String)
+                obj.put(trace[0], RSAHelper.decrypt(obj.getString(trace[0]), key));
         } else {
-            return null;
+            if (obj.get(trace[0]) instanceof JSONObject)
+                decrypt(Arrays.copyOfRange(trace, 1, trace.length), obj.getJSONObject(trace[0]), key);
         }
     }
 }