+++ CHANGELOG.md
@@ -1,9 +1,49 @@
 # 更新日记【Authz】
+
+## Version 1.0.9 - 2022.7.7
+
+### Added
+
+- 对于`@Decrypt` 新增了对象加密解密功能，支持对对象内某一个字段进行单独加密以及对整体加密，以及参数加密
+
+```java
+    @GetMapping("/get")
+public Result get(@Decrypt("name") String name){
+        return Result.SUCCESS.data("name",name);
+        }
+
+@PostMapping("/post")
+public Result post(@Decrypt({"name", "content", "obj.name"}) @RequestBody HashMap<String, Object> map){
+        return Result.SUCCESS.data("map",map);
+        }
+```
+
+- 若`@Decrypt`无参数传播，则key无限制,但值为整个加密的json，如
+
+```json
+{
+  "key名无限制": "value为整个json加密后的值，包含 `{` `}`"
+}
+```
+
+## Version 1.0.8 - 2022.7.5
+
+### Fixed
+
+- 修复了一些bug
+
+## Version 1.0.7 - 2022.7.5
+
+### Fixed
+
+- 修复了一些bug
+
 ## Version 1.0.6 - 2022.7.1
 
 ### Fixed
 
 - 修复了一些依赖bug
+- 修复了ObservableMap的兼容问题
 
 ## Version 1.0.5 - 2022.5.29